pyproject.toml
[tool.poetry]
name = "clipboard-manager"
version = "0.1.0"
description = "Clipboard image saver GUI for Windows"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "clipboard_manager" }]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.4.0"
pywin32 = "^306"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.0"

[tool.poetry.scripts]
clipboard-manager = "clipboard_manager.__main__:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

